Found conflicts between dependency naming of the headers and header guards and thus moved glfw and sdl to be named sisters_name convention to avoid conflicts
