+++ 3Sisters-Engine/inc/ecs/types/entity.hpp
@@ -10,6 +10,6 @@
 using Entity = std::uint32_t;
 
 // define the maximum amount of entites
-const Entity MAX_ENTITIES = 100000;
+const Entity MAX_ENTITIES = 10000;
 
 #endif
