+++ 3Sisters-Engine/src/input/managers/sisters_sdl_gamepad_manager.cpp
@@ -1,45 +1,35 @@
-#include <input/managers/glfw_gamepad_manager.hpp>
+#include <input/managers/sisters_sdl_gamepad_manager.hpp>
+
+#include <cstdlib>
+#include <SDL3/SDL.h>
 
 //? include IO for debug prints
 #include <iostream>
 
-using namespace GLFW;
+using namespace SDL;
 
 // define static variables
 std::vector<GamepadManager::QueuedGamepad>              GamepadManager::queuedGamepads;
 std::array<std::shared_ptr<ControllerDevice>, 15>       GamepadManager::devices;
+SDL_Event*                                              GamepadManager::eventHandle = nullptr;
 bool                                                    GamepadManager::isAutoClearSet = false;
 
-
-void GamepadManager::InitializeQuery(){
-    // check that GLFW has not been initialized
-    if(glfwGetError(NULL) == GLFW_NOT_INITIALIZED){
-        std::cout << "ERROR: GLFW is not initialized!\n";
-        return; // stop function
+void GamepadManager::InitializeQuery(SDL_Event& event){
+    // check if SDL hasn't been initialized
+    if(SDL_WasInit(SDL_INIT_GAMEPAD | SDL_INIT_JOYSTICK) != 0){
+        std::cout << "ERROR: SDL is not initialized!\n";
+        return; // stop functions 
     }
 
-    // check all slots for pre-connected devices
-    for(int i = 0; i < GLFW_JOYSTICK_LAST; i++){
-        // check if it is a gamepad and it's slot doesn't exist
-        if(glfwJoystickIsGamepad(i) == GLFW_TRUE && devices.at(i) == nullptr){
-            // create controller device
-            devices.at(i) = std::make_shared<ControllerDevice>();
-            // set the ID
-            devices.at(i)->ID = i;
-            // then set the connection to be true
-            devices.at(i)->isConnected = true;
-        }
-    }
-
-    // throw ID into a connection or disconnect events
-    glfwSetJoystickCallback(connectionCheck);
+    // set the event handle 
+    eventHandle = &event;
 }
 
 int GamepadManager::GetGamepadAmount(){
     // init var to contain actual amount of devices available
     int amount = 0;
 
-    for(int i = 0; i < GLFW_JOYSTICK_LAST; i++){
+    for(int i = 0; i < 15; i++){
         if(devices.at(i) != nullptr){
             amount++;
         }else{
@@ -55,12 +45,12 @@ void GamepadManager::SetGamepad(Gamepad& gamepad, int index){
     setUpAutoClear();
 
     // check index is within range and if reference already exists
-    if(index < GLFW_JOYSTICK_LAST && devices.at(index) != nullptr){
+    if(index < 15 && devices.at(index) != nullptr){
         // set the reference
         gamepad.device = devices.at(index);
 
         return; // stop function
-    }else if(index < GLFW_JOYSTICK_LAST && devices.at(index) == nullptr){
+    }else if(index < 15 && devices.at(index) == nullptr){
         // queue the given gamepad
         
         QueuedGamepad qp;
@@ -72,41 +62,45 @@ void GamepadManager::SetGamepad(Gamepad& gamepad, int index){
     }
 }
 
-void GamepadManager::PollInputs(){    
-    // check that GLFW has not been initialized
-    if(glfwGetError(NULL) == GLFW_NOT_INITIALIZED){
-        std::cout << "ERROR: GLFW is not initialized!\n";
-        return; // stop function
+void GamepadManager::PollIO(){
+    // check if the event handle is set meaning, there is intention of using the gamepad system
+    if(eventHandle == nullptr){
+        return; // stop functions
     }
 
-    // loop through available gamepads
-    for(int i = 0 ; devices.size(); i++){
-        if(devices.at(i) != nullptr && devices.at(i)->isConnected){
-            // update state
-            glfwGetGamepadState(i, &devices.at(i)->state);
-        }else if(devices.at(i) == nullptr){
-            return; // stop function 
-        }
+    // check if SDL and event handle hasn't been initialized
+    if(SDL_WasInit(SDL_INIT_GAMEPAD | SDL_INIT_JOYSTICK) != 0){
+        std::cout << "ERROR: SDL is not initialized!\n";
+        return; // stop functions
+    }
+
+    // check for events
+    switch(eventHandle->type){
+        case SDL_EVENT_GAMEPAD_ADDED:
+            // call for enable gamepad with the correspoding device
+            enableGamepad(eventHandle->cdevice.which);
+
+            break;
+        case SDL_EVENT_GAMEPAD_REMOVED:
+            // disable the gamepad that was removed
+            disableGamepad(eventHandle->cdevice.which);
+
+            break;
+        default:
+            break;
     }
 }
 
-void GamepadManager::connectionCheck(int jid, int event){
-    // check event type
-    if(event == GLFW_CONNECTED){
-        // create or enable a gamepad
-        enableGamepad(jid);
-    }else if(event == GLFW_DISCONNECTED){
-        // disable a gamepad
-        disableGamepad(jid);
-    }  
-}
 
 void GamepadManager::enableGamepad(int jid){
     // check if ID already exists
-    for(int i = 0; i < devices.size(); i++){
-        if(devices.at(i) != nullptr && devices.at(i)->ID == jid){
+    for(int i = 0; i < GetGamepadAmount(); i++){
+        if(devices.at(i)->ID == jid){
             // gamepad exists, then set the connection to be true
             devices.at(i)->isConnected = true;
+
+            // attempt to reopen
+            devices.at(jid)->controller = SDL_OpenGamepad(jid);
 
             // stop function
             return;
@@ -114,13 +108,15 @@ void GamepadManager::enableGamepad(int jid){
     }
 
     // if ID doesn't exist then create one
-    if(jid < GLFW_JOYSTICK_LAST){
+    if(jid < 15){
         // create a new gamepad
         devices.at(jid) = std::make_shared<ControllerDevice>();
         // set the ID
         devices.at(jid)->ID = jid;
         // then set the connection to be true
         devices.at(jid)->isConnected = true;
+        // then set the game controller
+        devices.at(jid)->controller = SDL_OpenGamepad(jid);
 
         // loop in reverse through queued up gamepads that need to be set
         for(int i = 0; i < queuedGamepads.size(); i++){
@@ -136,21 +132,32 @@ void GamepadManager::enableGamepad(int jid){
 }
 
 void GamepadManager::disableGamepad(int jid){
-    // check if ID already exists
-    if(devices.at(jid) != nullptr){
-        // turn off device
-        devices.at(jid)->isConnected = false;
+    // check for given ID which device was removed and properly close it
+    for(int i = 0; i < GetGamepadAmount(); i++){
+        // check which device matches the given id
+        if(SDL_GetJoystickID(SDL_GetGamepadJoystick(devices.at(i)->controller)) == jid){
+            // properly close the device
+            devices.at(i)->isConnected = false;
+            SDL_CloseGamepad(devices.at(i)->controller);
+
+            return; // stop function
+        }
     }
+    
 }
 
 void GamepadManager::clear(){
+    // de-allocate resources
+
     // remove all reference of queued gamepads
     for(QueuedGamepad& pad : queuedGamepads){
         pad.gamepad = nullptr;
     }
 
-    // clear the queued gamepad list
-    queuedGamepads.clear();
+    // close all created SDL joysticks
+    for(int i = 0; i < GetGamepadAmount(); i++){
+        SDL_CloseGamepad(devices.at(i)->controller);
+    }
 }
 
 void GamepadManager::setUpAutoClear(){
