+++ 3Sisters-Engine/src/input/sisters_glfw_keyboard.cpp
@@ -1,7 +1,7 @@
-#include <input/glfw_keyboard.hpp>
+#include <input/sisters_glfw_keyboard.hpp>
 
 // include Keyboard Manager
-#include <input/managers/glfw_keyboard_manager.hpp>
+#include <input/managers/sisters_glfw_keyboard_manager.hpp>
 
 using namespace GLFW;
