+++ 3Sisters-Engine/src/window/sisters_glfw_window.cpp
@@ -1,4 +1,4 @@
-#include <window/glfw_window.hpp>
+#include <window/sisters_glfw_window.hpp>
 
 // include the namespace
 using namespace GLFW;
